+++ Geometry_Wars/Geometry_Wars/Collision.h
@@ -3,5 +3,5 @@
 #include "PlayerController.h"
 #include "Bullet.h"
 
-bool HasCollided(Player player, int x, int y, int rayon);
-bool HasCollidedBullet(Bullet bullet, int x, int y, int rayon);
+bool HasCollided(Player player, float x, float y, float rayon);
+bool HasCollidedBullet(Bullet bullet, float x, float y, float rayon);
